Isolate quality comparison from engine facade
The quality metrics operation is self-contained, so it can move behind a dedicated module while preserving mcp_video.engine.compare_quality for client, server, CLI, and external callers. The module keeps PSNR/SSIM parsing behavior and narrows the engine facade further.

Constraint: Continue the approved engine decomposition without changing public import paths or tool signatures.

Rejected: Bundle compare_quality with stabilization/masking | those paths mutate output files and carry higher FFmpeg filter risk, so keeping this PR read-only lowers review conflict.

Confidence: high

Scope-risk: narrow

Directive: Keep compare_quality as a read-only analysis operation; do not mix it with output-producing transforms.

Tested: ruff check mcp_video/engine.py mcp_video/engine_compare_quality.py; ruff format --check mcp_video/engine.py mcp_video/engine_compare_quality.py; python3 -m pytest tests/test_engine_advanced.py -k 'compare_quality' -q --tb=short; python3 -m pytest tests/test_cli.py -k 'compare_quality' tests/test_server.py::TestVideoCompareQualityTool tests/test_red_team.py -k 'compare_quality' -q --tb=short; python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short

Not-tested: Full slow/real-media suite.
